Move and rename VectorizedVariable
