+++ docs/tutorials/observation_processes_measurements.qmd
@@ -24,9 +24,8 @@ from _tutorial_theme import theme_tutorial
 from pyrenew.observation import (
     Measurements,
     HierarchicalNormalNoise,
-    VectorizedRV,
 )
-from pyrenew.randomvariable import DistributionalVariable
+from pyrenew.randomvariable import DistributionalVariable, VectorizedVariable
 from pyrenew.deterministic import DeterministicVariable, DeterministicPMF
 ```
 
@@ -83,18 +82,18 @@ Measurement data typically exhibits **sensor-level variability**: different inst
 observed ~ Normal(predicted + sensor_mode[sensor], sensor_sd[sensor])
 ```
 
-The sensor-level RVs must implement `sample(n_groups=...)`. Use `VectorizedRV` to wrap simple distributions:
+The sensor-level RVs must implement `sample(n_groups=...)`. Use `VectorizedVariable` to wrap simple distributions:
 
 ```{python}
 # | label: noise-model-general
 # Sensor modes: zero-centered, allowing positive or negative bias
-sensor_mode_rv = VectorizedRV(
+sensor_mode_rv = VectorizedVariable(
     "vec_sensor_mode",
     DistributionalVariable("sensor_mode", dist.Normal(0, 0.5)),
 )
 
 # Sensor SDs: must be positive, truncated normal is a common choice
-sensor_sd_rv = VectorizedRV(
+sensor_sd_rv = VectorizedVariable(
     "vec_sensor_sd",
     DistributionalVariable(
         "sensor_sd", dist.TruncatedNormal(loc=0.3, scale=0.15, low=0.05)
@@ -355,13 +354,13 @@ For wastewater, a "sensor" is a WWTP/lab pair—the combination of treatment pla
 ```{python}
 # | label: ww-noise-model
 # Sensor-level mode: systematic differences between WWTP/lab pairs
-ww_sensor_mode_rv = VectorizedRV(
+ww_sensor_mode_rv = VectorizedVariable(
     "vec_ww_sensor_mode",
     DistributionalVariable("ww_sensor_mode", dist.Normal(0, 0.5)),
 )
 
 # Sensor-level SD: measurement variability within each WWTP/lab pair
-ww_sensor_sd_rv = VectorizedRV(
+ww_sensor_sd_rv = VectorizedVariable(
     "vec_ww_sensor_sd",
     DistributionalVariable(
         "ww_sensor_sd", dist.TruncatedNormal(loc=0.3, scale=0.15, low=0.10)
