+++ pyrenew/observation/__init__.py
@@ -28,7 +28,6 @@ from pyrenew.observation.noise import (
     MeasurementNoise,
     NegativeBinomialNoise,
     PoissonNoise,
-    VectorizedRV,
 )
 from pyrenew.observation.types import ObservationSample
 
@@ -44,7 +43,6 @@ __all__ = [
     "NegativeBinomialNoise",
     "MeasurementNoise",
     "HierarchicalNormalNoise",
-    "VectorizedRV",
     # Observation processes
     "Counts",
     "CountsBySubpop",
